feat(android): fourth Android correctness and stability pass
- Enforce num.clamp() int/double type conversions across all Dart models, widgets, and services
- Unify atomic lesson switching lifecycle, clearing stale UI state, AI tasks, and audio buffers
- Implement DirectStreamingResampler16k in AudioDecoder to halve peak decoding memory
- Fix WAV parser extensible subformat bounds, word alignment padding, and partial frame carryover
- Add C++ streaming UTF-8 accumulator and sanitize surrogate sequences in Llama bridge
- Implement terminal-aware LLM pre-emption and request-ID guarded cancellation in Llama and Whisper
- Remove artificial 500ms AudioService timeouts and serialize rapid audio load operations
- Normalize stale database processing transcript status on startup
- Add refresh on tab navigation and safe deletion of active lesson
- Add comprehensive Fourth Pass automated unit test suite (67/67 tests passing)

+++ app/test/unit/audio_segment_test.dart
@@ -87,8 +87,8 @@ void main() {
       int proposedStart = -500;
       int proposedEnd = 12000;
 
-      int validStart = proposedStart.clamp(0, totalDuration);
-      int validEnd = proposedEnd.clamp(0, totalDuration);
+      int validStart = proposedStart.clamp(0, totalDuration).toInt();
+      int validEnd = proposedEnd.clamp(0, totalDuration).toInt();
 
       expect(validStart, greaterThanOrEqualTo(0));
       expect(validEnd, lessThanOrEqualTo(totalDuration));
